backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0
requests>=2.32,<3.0
django-cors-headers>=4.4,<5.0
